cleanup: split subscription occurrence tests
